# Conflicts:
#	src/main/java/com/rentalsystem/swp/controllers/LoginController.java
#	src/main/java/com/rentalsystem/swp/controllers/ProfilePage.java

+++ src/main/java/com/rentalsystem/swp/controllers/EditItemController.java
@@ -6,10 +6,13 @@ import com.rentalsystem.swp.Repositories.UserRepository;
 import com.rentalsystem.swp.models.ItemProfile;
 import com.rentalsystem.swp.Repositories.ItemRepository;
 import com.rentalsystem.swp.models.UserProfile;
+import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 @SessionAttributes({"id", "loginData"})
@@ -24,16 +27,25 @@ public class EditItemController {
     }
 
     @RequestMapping(value = "/edititem", method = RequestMethod.GET)
-    public String showItem(@ModelAttribute("id") Integer id, Model model){
-        model.addAttribute("id", id);
-        ItemProfile itemProfile;
-        if (itemRepository.existsById(id)) {
-            itemProfile = itemRepository.findById(id).get();
+    public String showItem(@ModelAttribute("id") Integer id, Model model, HttpSession session) {
+
+        String test = (String) session.getAttribute("currentUser");
+
+        if (test == null){
+            model.addAttribute("auth", "false");
+            return "main";
         }
         else {
-            itemProfile = new ItemProfile("", "", "", 0, "", "da@yandex.ru");
+            UserProfile userProfile = userRepository.findByEmail(test).get();
+
+            ItemProfile itemProfile;
+            itemProfile = itemRepository.findById(id).get();
+
+            model.addAttribute("id", id);
+            model.addAttribute("itemProfile", itemProfile);
+            model.addAttribute("auth", "true");
+            model.addAttribute("userProfile", userProfile);
         }
-        model.addAttribute("itemProfile", itemProfile);
         return "edititem";
     }
 
@@ -41,30 +53,48 @@ public class EditItemController {
 
     @RequestMapping(value = "/delete", method = RequestMethod.GET)
     public String deleteItem(@ModelAttribute("id") Integer id,
-                             Model model) {
-        model.addAttribute("id", id);
-        UserProfile userProfile = userRepository.findByEmail(itemRepository.findById(id).get().getOwner()).get();
-        LoginData loginData = new LoginData();
-        loginData.setEmail(userProfile.getEmail());
-        loginData.setPassword(userProfile.getPassword());
-        model.addAttribute("loginData", loginData);
-        itemRepository.deleteById(id);
+                             Model model, HttpSession session) {
+
+        String test = (String) session.getAttribute("currentUser");
+
+        if (test == null) {
+            model.addAttribute("auth", "false");
+            return "main";
+        }
+        else {
+
+            UserProfile userProfile = userRepository.findByEmail(test).get();
+
+            model.addAttribute("id", id);
+            model.addAttribute("auth", "true");
+            model.addAttribute("userProfile", userProfile);
+
+            itemRepository.deleteById(id);
+        }
         return "forward:/profile";
     }
 
     @RequestMapping(value = "/edititem", method = RequestMethod.POST)
     public String editItem(@ModelAttribute("id") Integer id,
-            @ModelAttribute("itemProfile") ItemProfile itemProfile, Model model){
-        model.addAttribute("ItemProfile",  itemProfile);
-        itemRepository.deleteById(id);
-//        ItemProfile newItem = new ItemProfile();
-//        newItem.setName(item.getName());
-//        newItem.setDescription(item.getNametDescription());
-//        newItem.setTimeSlots(item.getTimeSlots());
-//        newItem.setCategory(item.getCategory()).;
+            @ModelAttribute("itemProfile") ItemProfile itemProfile, Model model,  HttpSession session) {
 
-        itemRepository.save(itemProfile);
-        return "login";
+        String test = (String) session.getAttribute("currentUser");
+
+        if (test == null) {
+            model.addAttribute("auth", "false");
+            return "main";
+        }
+        else {
+            UserProfile userProfile = userRepository.findByEmail(test).get();
+
+            model.addAttribute("auth", "true");
+            model.addAttribute("ItemProfile", itemProfile);
+            model.addAttribute("userProfile", userProfile);
+
+            itemRepository.deleteById(id);
+            itemRepository.save(itemProfile);
+        }
+        return "profile";
     }
 
 }
